chore: add .gitignore and .dockerignore
This files indicate what is ignored by git and by podman/docker.

It uses a ignore by default at the very beginning, and then only
accept what should be in the repository or container.

Be careful when adding new directories or new files in the base
directory of the repository, to consider if they impact some
change on this file.

The ignore by default policy minimize that some files could be
committed by mistake (logs, development credentials, etc...).

Signed-off-by: Alejandro Visiedo <alejandro.visiedo@gmail.com>